requirements.txt
numpy
pandas
seaborn
python-dotenv
pymysql
mysql-connector-python
scikit-learn
catboost
xgboost

# -e .
